Don't remove success files on rerun as ids will exist from the first run

+++ src/relion/zocalo/class2d_wrapper.py
@@ -124,7 +124,6 @@ class Class2DWrapper(zocalo.wrapper.BaseWrapper):
         if (job_dir / "RELION_JOB_EXIT_SUCCESS").exists():
             # This job over-writes a previous one
             job_is_rerun = True
-            (job_dir / "RELION_JOB_EXIT_SUCCESS").unlink()
         else:
             job_is_rerun = False
             job_dir.mkdir(parents=True, exist_ok=True)
@@ -195,7 +194,8 @@ class Class2DWrapper(zocalo.wrapper.BaseWrapper):
         result = subprocess.run(
             class2d_command, cwd=str(project_dir), capture_output=True
         )
-        (job_dir / "RELION_JOB_EXIT_SUCCESS").unlink()
+        if not job_is_rerun:
+            (job_dir / "RELION_JOB_EXIT_SUCCESS").unlink()
 
         # Register the Class2D job with the node creator
         self.log.info(f"Sending {job_type} to node creator")
@@ -364,6 +364,6 @@ class Class2DWrapper(zocalo.wrapper.BaseWrapper):
             }
             self.recwrap.send_to("murfey_feedback", murfey_params)
 
-        (job_dir / "RELION_JOB_EXIT_SUCCESS").touch()
+        (job_dir / "RELION_JOB_EXIT_SUCCESS").touch(exist_ok=True)
         self.log.info(f"Done {job_type} for {class2d_params.particles_file}.")
         return True
