+++ src/relion/zocalo/class3d_wrapper.py
@@ -321,7 +321,6 @@ class Class3DWrapper(zocalo.wrapper.BaseWrapper):
         if (job_dir / "RELION_JOB_EXIT_SUCCESS").exists():
             # This job over-writes a previous one
             job_is_rerun = True
-            (job_dir / "RELION_JOB_EXIT_SUCCESS").unlink()
         else:
             job_is_rerun = False
             job_dir.mkdir(parents=True, exist_ok=True)
@@ -398,7 +397,8 @@ class Class3DWrapper(zocalo.wrapper.BaseWrapper):
         result = subprocess.run(
             class3d_command, cwd=str(project_dir), capture_output=True
         )
-        (job_dir / "RELION_JOB_EXIT_SUCCESS").unlink()
+        if not job_is_rerun:
+            (job_dir / "RELION_JOB_EXIT_SUCCESS").unlink()
 
         # Register the Class3D job with the node creator
         self.log.info(f"Sending {job_type} to node creator")
@@ -523,6 +523,6 @@ class Class3DWrapper(zocalo.wrapper.BaseWrapper):
         }
         self.recwrap.send_to("murfey_feedback", murfey_params)
 
-        (job_dir / "RELION_JOB_EXIT_SUCCESS").touch()
+        (job_dir / "RELION_JOB_EXIT_SUCCESS").touch(exist_ok=True)
         self.log.info(f"Done {job_type} for {class3d_params.particles_file}.")
         return True
